Use font hash in the key for glyph positions
The key slot of fontsloth-layout-glyph-position is intended for use as
a hash key. This replaces font layout index with font hash as part of
the key in order to ensure its uniqueness for different fonts.

Since this requires adding a hash slot to fontsloth-font, this change
also includes cache invalidation logic when reading from
fontsloth-cache so that stale cache entries are detected.

A new slot, -data-version, is added to the beginning of the
fontsloth-font record. Cache lookups now check that data in cache has
a value in that position matching the expected value.