Outliner: use a fixed object pool for curve segments
The segment stack for curve-to and quad-to is now backed by a fixed
vector of segments. This dramatically reduces the amount of garbage
generated during outlining since before new memory was allocated for
every curve segment of every glyph.

The initial size is controlled by a variable,
`fontsloth-seg-stack--size'.

To fontsloth-geometry, this adds one new type, `fontsloth-seg-stack'.

To operate on the new type are four new functions:
- fontsloth-seg-stack--reset
- fontsloth-seg-stack--grow
- fontsloth-seg-stack--push
- fontsloth-seg-stack--pop